requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --keyring-provider subprocess requirements.in -o requirements.txt
pillow==12.1.1
    # via qrcode
python-dotenv==1.2.1
    # via -r requirements.in
qrcode==8.2
    # via -r requirements.in
ruff==0.15.0
    # via -r requirements.in
uv==0.10.2
    # via -r requirements.in
